Add index size metric to vector store metrics
This patch adds an index size metric to the vector store metrics.
The engine now receives a metrics object and passes it to monitor_items.

It adds a dirty set to Metrics to record indexes that have changed.
Indexes are marked dirty on update, and on Prometheus scrape, the set is
collected and cleared and size is updated by calling count(). This
avoids redundant metric updates and ensures accuracy.

Whenever monitor_items updates the index, after completing the update,
it marks the index as dirty.

Example metric output:

  # HELP index_size Number of vectors per index
  # TYPE index_size gauge
  index_size{index_name="myindex",keyspace="mykey"} 99

Signed-off-by: Amnon Heiman <amnon@scylladb.com>