Temporary use zero UUID as an index version
We're waiting for the VECTOR-142 issue, the issue will bring solution to
assigning unique version to each index creation. We've got this version from
the Scylla view version, which is not accessible anymore. This patch gives a
workaround and creates a zero UUID for each index. We won't discover a fast
drop/create index with the same parameters, but it seems not urgent an edge
case (we simply didn't recreate index in that situation).