Implement script for building release binary
Currently the Dockerfile doesn't check rust toolchain version and do not use
rust release toolchain. This patch removes Dockerfile and creates a docker from
the script using `scylladb/vector-store:{version}` as a tag. The script will
rebuild the binary of vector-store service.

The script will prepare a release version of vector-store in a
`vector-store-{version}.tar.gz` tarball in a `target/release` directory.

This is a first step to prepare automation for releasing.