vector-store: implement usearch simulator
The commit implements simulator mode for usearch index. The simulator is
helpful when you need to test whole add/ann pipeline with simulation of usearch
work without running it (it consumes a lot of memory resources).

Env variable VECTOR_STORE_USEARCH_SIMULATES provides parameters for it: it is a
list of human readable durations, delimiter is ':' with such items:

- search duration
- add or remove duration
- reserver duration

The simulator ignores vector values, search results are random. The simulator
waits for specified time for reserve/add/remove/search operations - it
simulates time which usearch will spend on each operation.