vector-store: refactor index usearch with memory actor usage
During testing there were a lot of logs for `Unable to add vector...` when a
memory limit was reached. To limit logs and display them only on the allocation
status changes there was a need for refactoring usage of the memory actor by
the usearch index. Now condition is checked before spawning a new
add_or_replace task.