validator: fix total test number statistic
When running Validator with subset of tests, the `total`
number of tests in statistics was incorrect.
It represented all tests in a test file, not only those selected.

This fix changes the `total` number of tests to only the tests
being run. The pass results should show `total` = `launched` = `ok`.

Fixes: VECTOR-392